requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.0.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
CacheControl==0.14.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==42.0.7
databases==0.9.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.12.2
firebase-admin==6.5.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
identify==2.5.25
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.4
nodeenv==1.8.0
orjson==3.10.3
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
pre-commit==3.3.3
proto-plus==1.23.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==4.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.29.0
uvloop==0.19.0
validators==0.20.0
virtualenv==20.24.1
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zipp==3.15.0
